change int to float in download airway bill price

+++ my_app/views/locations/views.py
@@ -86,22 +86,15 @@ class AirwayBillLocationDeleteView(View):
         def post(self, request, *args, **kwargs):
             return self._get(request, *args, **kwargs)
 
-# class DownloadAirwayBillView(View):
 
-#     def get(self, request, bill_id):
-#         bill = AirwayBill.objects.get(id=bill_id)
-#         context_dict = {
-#             'bill':bill
-#         }
-#         html_content = render_to_string('billing_locations/download_billings.html', context_dict)
-#         return JsonResponse({'html_content': html_content})
-        
-#         # return render(request,template_name='billing_locations/download_billings.html',context=context_dict)
+
 class DownloadAirwayBillView(View):
 
     def get(self, request, bill_id):
         bill = AirwayBill.objects.get(id=bill_id)
-        total_price = sum(int(detail.get('price')) for detail in bill.data.get('invoice_details').values())
+        # total_price = sum(int(detail.get('price')) for detail in bill.data.get('invoice_details').values())
+        total_price = sum(float(detail.get('price')) for detail in bill.data.get('invoice_details').values())
+
         context_dict = {
             'bill':bill,
             'total_price':total_price
